Add access checks for PurchaseDetail entity. Rename IsPurchaseOwnerOrShopManagerAsync to IsPurchaseOwnerOrShopManagementAsync and make it so it check for shopowner as well as shop manager

+++ AssistCust.Application/PurchasesDetails/Commands/DeletePurchaseDetail/DeletePurchaseDetailCommandHandler.cs
@@ -10,18 +10,24 @@ namespace AssistCust.Application.PurchaseDetails.Commands.DeletePurchaseDetail
     public class DeletePurchaseDetailCommandCommandHandler : IRequestHandler<DeletePurchaseDetailCommand, Unit>
     {
         private readonly IAssistDbContext _context;
-        public DeletePurchaseDetailCommandCommandHandler(IAssistDbContext context)
+        private readonly IUserAccessService _userAccessService;
+        public DeletePurchaseDetailCommandCommandHandler(IAssistDbContext context, IUserAccessService userAccessService)
         {
             _context = context;
+            _userAccessService = userAccessService;
         }
         public async Task<Unit> Handle(DeletePurchaseDetailCommand request, CancellationToken cancellationToken)
         {
+
             var entity = await _context.PurchaseDetails.FindAsync(request.Id);
 
             if (entity == null)
             {
                 throw new NotFoundException(nameof(PurchaseDetail), request.Id);
             }
+
+            if (!(await _userAccessService.IsPurchaseOwnerOrShopManagementAsync(entity.PurchaseId)))
+                throw new InsufficientPrivilegesException(nameof(PurchaseDetail));
 
             _context.PurchaseDetails.Remove(entity);
